requirements.txt
ansible_runner==2.1.4
natsort==8.0.2
PyYAML==6.0
ujson==5.4.0
urllib3==1.26.19
validators==0.22.0
jinja2==2.10.1
setuptools==70.0.0
